perf: 资产协议组自治，关键时刻不跟随平台协议联动
